% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{create_items}
\alias{create_items}
\title{Create todo items}
\usage{
create_items(conn, todo, items)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}

\item{items}{A character vector of the new todo items}
}
\description{
Create todo items
}
