% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{create_todo}
\alias{create_todo}
\title{Create todo list}
\usage{
create_todo(conn, todo)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}
}
\description{
Creates a new todo list (sqlite table) with the given name. Uses
\href{https://solutions.posit.co/connections/db/best-practices/run-queries-safely/#interpolation-by-hand}{interpolation by hand}
}
