% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_all_items}
\alias{delete_all_items}
\title{Delete all items in a todo list}
\usage{
delete_all_items(conn, todo)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}
}
\description{
Delete all items in a todo list
}
