% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_items}
\alias{delete_items}
\title{Delete todo items}
\usage{
delete_items(conn, todo, item_ids)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}

\item{item_ids}{Ids of the items to delete}
}
\description{
Delete todo items
}
