% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{delete_todo}
\alias{delete_todo}
\title{Deletes a todo list}
\usage{
delete_todo(conn, todo)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}
}
\description{
Deletes a todo list
}
