% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{read_item}
\alias{read_item}
\title{Read a specific item in a todo list}
\usage{
read_item(conn, todo, item_id)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}

\item{item_id}{Item id}
}
\description{
Read a specific item in a todo list
}
