% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{read_items}
\alias{read_items}
\title{Read items of a todo list}
\usage{
read_items(conn, todo)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}
}
\description{
Read items of a todo list
}
