% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{update_item}
\alias{update_item}
\title{Update a todo list item}
\usage{
update_item(conn, todo, item_id, updated_item)
}
\arguments{
\item{conn}{DB connection}

\item{todo}{Name of the todo list}

\item{item_id}{Item id}

\item{updated_item}{Updated item}
}
\description{
Update a todo list item
}
